back/requirements.txt
flask
python-dotenv
mysql-connector-python
flask-bcrypt
